backend/requirements.txt
flask==3.0.0
werkzeug==3.0.1
numpy==1.24.3
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==1.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
price-parser==0.3.4
python-dateutil==2.8.2
